rename HaztrakSite directory to Site directory
