+++ client/app/components/Manifest/SiteSelect/SiteSelect.tsx
@@ -3,7 +3,7 @@ import { HtForm } from 'app/components/legacyUi';
 import React, { useMemo } from 'react';
 import { Control, Controller } from 'react-hook-form';
 import Select from 'react-select';
-import { HaztrakSite } from '~/components/HaztrakSite';
+import { HaztrakSite } from 'app/components/Site';
 import { RcraSite } from '~/components/RcraSite';
 import { useGetUserHaztrakSitesQuery } from '~/store';
