+++ client/app/components/Site/index.ts
@@ -1,4 +1,4 @@
 export { SiteListGroup } from './SiteListGroup/SiteListGroup';
-import { HaztrakSite } from './haztrakSiteSchema';
+import { HaztrakSite } from 'app/components/Site/siteSchema';
 
 export type { HaztrakSite };
